+++ src/components/pages/index.js
@@ -1,6 +1,5 @@
 import Appointment from "./Appointment";
 import CheckIn from "./CheckIn";
 import Welcome from "./Welcome";
-import Layout from "./Layout";
 
-export { Appointment, CheckIn, Welcome, Layout };
+export { Appointment, CheckIn, Welcome };
