--- src/index.css
@@ -1,5 +1,0 @@
-html,
-body,
-#root {
-  height: 100%;
-}
