+++ src/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
+import App from "./components/App";
 import { BrowserRouter as Router } from "react-router-dom";
 
 ReactDOM.render(
